Fixed server.js
