+++ src/index.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <!-- <link rel="stylesheet" type="text/css" href="style.css" /> -->
     <title>GoodReads</title>
+    <script async defer src="bundle.js"></script>
   </head>
   <body>
     <div id="root"></div>
